Extract lead intent and match into leads-core
Ranking buying intent, and recognising our own drafted comment once a
human has pasted it into a thread, are the two pieces of this product a
lead-finding vertical would need on its own. Both were already pure
functions, but they lived inside the app and were reachable only through
$lib, so nothing outside SvelteKit could use them.

They move to `@anomalia/leads-core`, the first package of the leads
core. `packages/no-app-imports.test.ts` already guards it: a $lib or
$env import there fails the build from now on.

Nothing changes behaviourally — same functions, same tests, a different
address. radar.ts also loses its re-export of the intent symbols, which
no caller ever used.

The IO half of lead-outcomes stays in the app: it needs the Supabase
client, ScrapeCreators and the app's Sentry-backed error reporter, and
inverting those is a separate move.
